inserting the picture
Subfolders assets, styles, and scripts has been created. The example picture has been added to jumbotron container and its size has been set to 100px/100px. That is for sure too small. To do: work with picture - to make it to be centered, bigger (if on big screen), and responsice. Start with the mobile view.